# Conflicts:
#	VolleyManagement.UI/App_GlobalResources/TeamViews.Designer.cs
#	VolleyManagement.UI/App_GlobalResources/ViewModelResources.Designer.cs
#	VolleyManagement.UI/App_Start/BundleConfig.cs
#	VolleyManagement.UI/Areas/Mvc/Controllers/Domain/PlayersController.cs
#	VolleyManagement.UI/Areas/Mvc/Views/Shared/_Layout.cshtml
#	VolleyManagement.UI/Areas/Mvc/Views/Teams/Index.cshtml
#	VolleyManagement.UI/VolleyManagement.UI.csproj

+++ VolleyManagement.UI/Areas/Mvc/Controllers/Domain/TournamentsController.cs
@@ -5,19 +5,22 @@
     using System.Web.Mvc;
     using VolleyManagement.Contracts;
     using VolleyManagement.Contracts.Exceptions;
-    using VolleyManagement.Domain.Tournaments;
+    using VolleyManagement.Domain;
     using VolleyManagement.Domain.TournamentsAggregate;
-    using VolleyManagement.UI.Areas.Mvc.Mappers;
     using VolleyManagement.UI.Areas.Mvc.ViewModels.Tournaments;
-
-    using ErrorMessages = VolleyManagement.Domain.Properties.Resources;
-    using ValidationMessages = App_GlobalResources.ViewModelResources;
 
     /// <summary>
     /// Defines TournamentsController
     /// </summary>
     public class TournamentsController : Controller
     {
+        private const int DAYS_TO_APPLYING_PERIOD_START = 1;
+        private const int DAYS_FOR_APPLYING_PERIOD = Constants.Tournament.DAYS_BETWEEN_START_AND_END_APPLYING_DATE;
+        private const int DAYS_FROM_APPLYING_PERIOD_END_TO_GAMES_START = 1;
+        private const int DAYS_FOR_GAMES_PERIOD = 120;
+        private const int DAYS_FROM_GAMES_START_TO_TRANSFER_START = 1;
+        private const int DAYS_FOR_TRANSFER_PERIOD = 20;
+
         /// <summary>
         /// Holds TournamentService instance
         /// </summary>
@@ -69,15 +72,7 @@
         /// <returns>View with specific tournament</returns>
         public ActionResult Details(int id)
         {
-            try
-            {
-                Tournament tournament = this._tournamentService.Get(id);
-                return View(tournament);
-            }
-            catch (InvalidOperationException)
-            {
-                return this.HttpNotFound();
-            }
+            return GetTournamentView(id);
         }
 
         /// <summary>
@@ -87,14 +82,18 @@
         public ActionResult Create()
         {
             var tournamentViewModel = new TournamentViewModel()
-                {
-                    ApplyingPeriodStart = DateTime.Now.AddDays(1),
-                    ApplyingPeriodEnd = DateTime.Now.AddDays(1),
-                    GamesStart = DateTime.Now.AddDays(1),
-                    GamesEnd = DateTime.Now.AddDays(1),
-                    TransferStart = DateTime.Now.AddDays(1),
-                    TransferEnd = DateTime.Now.AddDays(1)
-                };
+            {
+                ApplyingPeriodStart = DateTime.Now.AddDays(DAYS_TO_APPLYING_PERIOD_START),
+                ApplyingPeriodEnd = DateTime.Now.AddDays(DAYS_TO_APPLYING_PERIOD_START + DAYS_FOR_APPLYING_PERIOD),
+                GamesStart = DateTime.Now.AddDays(DAYS_TO_APPLYING_PERIOD_START + DAYS_FOR_APPLYING_PERIOD
+                + DAYS_FROM_APPLYING_PERIOD_END_TO_GAMES_START),
+                GamesEnd = DateTime.Now.AddDays(DAYS_TO_APPLYING_PERIOD_START + DAYS_FOR_APPLYING_PERIOD
+                + DAYS_FROM_APPLYING_PERIOD_END_TO_GAMES_START + DAYS_FOR_GAMES_PERIOD),
+                TransferStart = DateTime.Now.AddDays(DAYS_TO_APPLYING_PERIOD_START + DAYS_FOR_APPLYING_PERIOD
+                + DAYS_FROM_APPLYING_PERIOD_END_TO_GAMES_START + DAYS_FROM_GAMES_START_TO_TRANSFER_START),
+                TransferEnd = DateTime.Now.AddDays(DAYS_TO_APPLYING_PERIOD_START + DAYS_FOR_APPLYING_PERIOD
+                + DAYS_FROM_APPLYING_PERIOD_END_TO_GAMES_START + DAYS_FROM_GAMES_START_TO_TRANSFER_START + DAYS_FOR_TRANSFER_PERIOD)
+            };
 
             return this.View(tournamentViewModel);
         }
@@ -123,10 +122,6 @@
                 this.ModelState.AddModelError(e.ValidationKey, e.Message);
                 return this.View(tournamentViewModel);
             }
-            catch (Exception)
-            {
-                return this.HttpNotFound();
-            }
         }
 
         /// <summary>
@@ -136,16 +131,7 @@
         /// <returns>View to edit specific tournament</returns>
         public ActionResult Edit(int id)
         {
-            try
-            {
-                var tournament = this._tournamentService.Get(id);
-                TournamentViewModel tournamentViewModel = TournamentViewModel.Map(tournament);
-                return this.View(tournamentViewModel);
-            }
-            catch (Exception)
-            {
-                return this.HttpNotFound();
-            }
+            return GetTournamentView(id);
         }
 
         /// <summary>
@@ -172,10 +158,6 @@
                 this.ModelState.AddModelError(string.Empty, ex.Message);
                 return this.View(tournamentViewModel);
             }
-            catch (Exception)
-            {
-                return this.HttpNotFound();
-            }
         }
 
         /// <summary>
@@ -185,9 +167,7 @@
         /// <returns>View to delete specific tournament</returns>
         public ActionResult Delete(int id)
         {
-            Tournament tournament = this._tournamentService.Get(id);
-            TournamentViewModel tournamentViewModel = TournamentViewModel.Map(tournament);
-            return View(tournament);
+            return GetTournamentView(id);
         }
 
         /// <summary>
@@ -198,24 +178,33 @@
         [HttpPost, ActionName("Delete")]
         public ActionResult DeleteConfirmed(int id)
         {
-            // This will return "An error occured in VolleyManagement application."
-            // Please contact site administrator." if tournament doesnt exist
-            // this._tournamentService.Delete(id);
-            // return this.RedirectToAction("Index");
+            var tournament = _tournamentService.Get(id);
 
-            // This will return 404
-            ActionResult result;
-            try
+            if (tournament == null)
             {
-                this._tournamentService.Delete(id);
-                result = this.RedirectToAction("Index");
-            }
-            catch (Exception)
-            {
-                result = this.HttpNotFound();
+                return HttpNotFound();
             }
 
-            return result;
+            this._tournamentService.Delete(id);
+            return RedirectToAction("Index");
+        }
+
+        /// <summary>
+        /// Gets the view for view model of the tournament with specified identifier.
+        /// </summary>
+        /// <param name="id">Identifier of the tournament.</param>
+        /// <returns>View for view model of the tournament with specified identifier.</returns>
+        private ActionResult GetTournamentView(int id)
+        {
+            var tournament = _tournamentService.Get(id);
+
+            if (tournament == null)
+            {
+                return HttpNotFound();
+            }
+
+            var tournamentViewModel = TournamentViewModel.Map(tournament);
+            return View(tournamentViewModel);
         }
     }
 }
